# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "AnswerOut": {
        "properties": {
          "conclusive": {
            "title": "Conclusive",
            "type": "boolean"
          },
          "elapsed_s": {
            "title": "Elapsed S",
            "type": "number"
          },
          "focus": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FocusOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "From the tools called, never from the model"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "sources": {
            "description": "Tools actually read, not what the model claims",
            "items": {
              "type": "string"
            },
            "title": "Sources",
            "type": "array"
          },
          "stance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stance"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/StepOut"
            },
            "title": "Steps",
            "type": "array"
          },
          "structured": {
            "title": "Structured",
            "type": "boolean"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "ungrounded": {
            "description": "Figures in the text that no tool returned",
            "items": {
              "type": "string"
            },
            "title": "Ungrounded",
            "type": "array"
          }
        },
        "required": [
          "text",
          "conclusive",
          "steps",
          "sources",
          "ungrounded",
          "provider",
          "model",
          "elapsed_s",
          "structured",
          "focus"
        ],
        "title": "AnswerOut",
        "type": "object"
      },
...
      "AskIn": {
        "properties": {
          "context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContextIn"
              },
              {
                "type": "null"
              }
            ]
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/TurnIn"
            },
            "maxItems": 12,
            "title": "History",
            "type": "array"
          },
          "question": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "AskIn",
        "type": "object"
      },
...
      "ContextIn": {
        "properties": {
          "details": {
            "additionalProperties": {
              "type": "string"
            },
            "maxProperties": 20,
            "title": "Details",
            "type": "object"
          },
          "kind": {
            "maxLength": 60,
            "title": "Kind",
            "type": "string"
          },
          "lat": {
            "anyOf": [
              {
                "maximum": 90.0,
                "minimum": -90.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "lon": {
            "anyOf": [
              {
                "maximum": 180.0,
                "minimum": -180.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lon"
          },
          "title": {
            "maxLength": 200,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "title"
        ],
        "title": "ContextIn",
        "type": "object"
      },
...
      "DocKind": {
        "enum": [
          "story",
          "post"
        ],
        "title": "DocKind",
        "type": "string"
      },
      "Document": {
        "properties": {
          "at": {
            "format": "date-time",
            "title": "At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/DocKind"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "text": {
            "default": "",
            "title": "Text",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "unverified": {
            "description": "Telegram posts and state-media-only stories",
            "title": "Unverified",
            "type": "boolean"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "title",
          "text",
          "source",
          "url",
          "at",
          "unverified"
        ],
        "title": "Document",
        "type": "object"
      },
...
      "FocusOut": {
        "properties": {
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "layers": {
            "items": {
              "type": "string"
            },
            "title": "Layers",
            "type": "array"
          },
          "lon": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lon"
          },
          "zoom": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zoom"
          }
        },
        "required": [
          "lat",
          "lon",
          "zoom",
          "country",
          "layers"
        ],
        "title": "FocusOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Hit": {
        "properties": {
          "document": {
            "$ref": "#/components/schemas/Document"
          },
          "match": {
            "$ref": "#/components/schemas/Match"
          },
          "score": {
            "title": "Score",
            "type": "number"
          }
        },
        "required": [
          "document",
          "score",
          "match"
        ],
        "title": "Hit",
        "type": "object"
      },
...
      "Match": {
        "enum": [
          "lexical",
          "semantic",
          "both"
        ],
        "title": "Match",
        "type": "string"
      },
...
      "ModelUsage": {
        "properties": {
          "calls": {
            "title": "Calls",
            "type": "integer"
          },
          "calls_without_cost": {
            "description": "Calls whose price LiteLLM does not know",
            "title": "Calls Without Cost",
            "type": "integer"
          },
          "cost_usd": {
            "title": "Cost Usd",
            "type": "number"
          },
          "input_tokens": {
            "title": "Input Tokens",
            "type": "integer"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "output_tokens": {
            "title": "Output Tokens",
            "type": "integer"
          }
        },
        "required": [
          "model",
          "calls",
          "input_tokens",
          "output_tokens",
          "cost_usd",
          "calls_without_cost"
        ],
        "title": "ModelUsage",
        "type": "object"
      },
...
      "PredictionIn": {
        "properties": {
          "market_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Market Id",
            "type": "string"
          },
          "tag": {
            "default": "geopolitics",
            "pattern": "^[a-z0-9-]+$",
            "title": "Tag",
            "type": "string"
          }
        },
        "required": [
          "market_id"
        ],
        "title": "PredictionIn",
        "type": "object"
      },
...
      "PredictionReadingOut": {
        "properties": {
          "answer": {
            "$ref": "#/components/schemas/AnswerOut"
          },
          "market_id": {
            "title": "Market Id",
            "type": "string"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "stance": {
            "enum": [
              "leans_yes",
              "leans_no",
              "no_conclusion"
            ],
            "title": "Stance",
            "type": "string"
          }
        },
        "required": [
          "market_id",
          "question",
          "stance",
          "answer"
        ],
        "title": "PredictionReadingOut",
        "type": "object"
      },
...
      "SearchOut": {
        "properties": {
          "hits": {
            "items": {
              "$ref": "#/components/schemas/Hit"
            },
            "title": "Hits",
            "type": "array"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "searched": {
            "title": "Searched",
            "type": "integer"
          },
          "semantic": {
            "description": "False: keywords only (no embedding model configured)",
            "title": "Semantic",
            "type": "boolean"
          }
        },
        "required": [
          "query",
          "hits",
          "semantic",
          "searched",
          "note"
        ],
        "title": "SearchOut",
        "type": "object"
      },
...
      "SettingsIn": {
        "properties": {
          "api_base": {
            "anyOf": [
              {
                "maxLength": 500,
                "pattern": "^https?://\\S+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Base"
          },
          "api_key": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Omit to keep the stored key; empty to remove it",
            "title": "Api Key"
          },
          "embedding_model": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Model"
          },
          "model": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          }
        },
        "title": "SettingsIn",
        "type": "object"
      },
      "SettingsOut": {
        "properties": {
          "api_base": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Base"
          },
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "embedding_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Semantic search; None = keywords only",
            "title": "Embedding Model"
          },
          "fallbacks": {
            "items": {
              "type": "string"
            },
            "title": "Fallbacks",
            "type": "array"
          },
          "key_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Last characters only; the key is never returned",
            "title": "Key Hint"
          },
          "key_set": {
            "title": "Key Set",
            "type": "boolean"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "source": {
            "enum": [
              "app",
              "environment",
              "none"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "model",
          "api_base",
          "key_set",
          "key_hint",
          "source",
          "fallbacks",
          "available",
          "embedding_model"
        ],
        "title": "SettingsOut",
        "type": "object"
      },
...
      "StepOut": {
        "properties": {
          "args": {
            "additionalProperties": true,
            "title": "Args",
            "type": "object"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "tool": {
            "title": "Tool",
            "type": "string"
          }
        },
        "required": [
          "tool",
          "args",
          "ok",
          "summary"
        ],
        "title": "StepOut",
        "type": "object"
      },
...
      "TurnIn": {
        "properties": {
          "content": {
            "maxLength": 4000,
            "title": "Content",
            "type": "string"
          },
          "role": {
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "content"
        ],
        "title": "TurnIn",
        "type": "object"
      },
      "UsageOut": {
        "properties": {
          "by_model": {
            "items": {
              "$ref": "#/components/schemas/ModelUsage"
            },
            "title": "By Model",
            "type": "array"
          },
          "calls": {
            "title": "Calls",
            "type": "integer"
          },
          "calls_without_cost": {
            "title": "Calls Without Cost",
            "type": "integer"
          },
          "cost_usd": {
            "description": "LiteLLM's estimate; local models count as 0",
            "title": "Cost Usd",
            "type": "number"
          },
          "days": {
            "title": "Days",
            "type": "integer"
          },
          "input_tokens": {
            "title": "Input Tokens",
            "type": "integer"
          },
          "output_tokens": {
            "title": "Output Tokens",
            "type": "integer"
          }
        },
        "required": [
          "days",
          "calls",
          "input_tokens",
          "output_tokens",
          "cost_usd",
          "calls_without_cost",
          "by_model"
        ],
        "title": "UsageOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Argus API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/assistant/analyse/prediction": {
      "post": {
        "description": "Whether Argus' own data leans the way a prediction market prices \u2014 or not at all.",
        "operationId": "analyse_prediction_api_v1_assistant_analyse_prediction_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionReadingOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyse Prediction",
        "tags": [
          "assistant"
        ]
      }
    },
    "/api/v1/assistant/ask": {
      "post": {
        "description": "Answer from Argus' own data: the steps, the sources read, and unbacked figures.",
        "operationId": "ask_api_v1_assistant_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ask",
        "tags": [
          "assistant"
        ]
      }
    },
    "/api/v1/assistant/models": {
      "get": {
        "description": "Models pulled on an Ollama server, to pick from.",
        "operationId": "local_models_api_v1_assistant_models_get",
        "parameters": [
          {
            "in": "query",
            "name": "api_base",
            "required": true,
            "schema": {
              "maxLength": 500,
              "pattern": "^https?://\\S+$",
              "title": "Api Base",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Local Models Api V1 Assistant Models Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Local Models",
        "tags": [
          "assistant"
        ]
      }
    },
...
    "/api/v1/assistant/search": {
      "get": {
        "description": "Hybrid search over recent news stories and the owner's Telegram channels.",
        "operationId": "search_api_v1_assistant_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 2,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "hours",
            "required": false,
            "schema": {
              "default": 72,
              "maximum": 168,
              "minimum": 1,
              "title": "Hours",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 12,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search",
        "tags": [
          "assistant"
        ]
      }
    },
    "/api/v1/assistant/settings": {
      "get": {
        "operationId": "get_settings_api_v1_assistant_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Settings",
        "tags": [
          "assistant"
        ]
      },
      "put": {
        "description": "Choose the model. The key is stored server side and never sent back.",
        "operationId": "put_settings_api_v1_assistant_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Settings",
        "tags": [
          "assistant"
        ]
      }
    },
...
    "/api/v1/assistant/usage": {
      "get": {
        "description": "Model calls, tokens and estimated cost over the last `days`.",
        "operationId": "usage_api_v1_assistant_usage_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 365,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Usage",
        "tags": [
          "assistant"
        ]
      }
    },
...
  }
}